requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
Cython==0.29.33
dj-database-url==2.2.0
Django==5.0.7
django-admin-honeypot-updated-2021==1.2.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-ipware==7.0.1
django-resized==1.0.2
g==0.0.5
gunicorn==23.0.0
idna==3.8
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-ipware==3.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
